# spec/openapi.yaml
swagger: '2.0'
info:
  title: AI Gateway
  version: 1.0.0
  description: |
    A platform for machine learning & high performance computing
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
host: 'localhost:9000'
basePath: /api/v1
securityDefinitions:
  api-authorizer:
    name: Authorization
    type: apiKey
    in: header
paths:
...
  '/remote-images/{id}':
    get:
      description: |
        returns remote images
      operationId: getRemoteImages
      tags:
        - repository
      security:
        - api-authorizer: []
      parameters:
        - name: id
          in: path
          description: Docker image name
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Image'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/nvidia/repositories':
    get:
      description: |
        returns NGC repositories
      operationId: getNgcRepositories
      tags:
        - repository
      security:
        - api-authorizer: []
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Repository'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/nvidia/repositories/{namespace}/images/{id}':
    get:
      description: |
        returns NGC images
      operationId: getNgcImages
      tags:
        - repository
      security:
        - api-authorizer: []
      parameters:
        - name: namespace
          in: path
          description: Docker repositry namespace
          type: string
          required: true
        - name: id
          in: path
          description: Docker image name
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/NgcImage'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/notebooks':
    get:
      description: |
        returns Jupyter notebook information
      operationId: getNotebooks
      tags:
        - notebook
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Notebook'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: |
        creates Jupyter notebook container
      operationId: postNewNotebook
      tags:
        - notebook
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: ImageName
            type: object
            required:
              - image
            properties:
              image:
                description: Docker image name
                type: string
              workspace:
                description: workspace to be mounted
                type: string
      responses:
        '201':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/notebooks/{id}':
    get:
      description: |
        returns Jupyter notebook detail information
      operationId: getNotebookDetails
      tags:
        - notebook
      parameters:
        - name: id
          in: path
          description: Notebook container ID
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/NotebookDetail'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    patch:
      description: |
        modify the notebook status
      operationId: modifyNotebook
      tags:
        - notebook
      parameters:
        - name: id
          in: path
          description: Notebook container ID
          type: string
          required: true
        - name: body
          in: body
          required: true
          schema:
            title: NotebookAttrs
            type: object
            properties:
              status:
                description: status
                type: string
                enum:
                    - started
                    - stopped
      responses:
        '200':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: |
        delete a specified notebook
      operationId: deleteNotebook
      tags:
        - notebook
      parameters:
        - name: id
          in: path
          description: Notebook container ID
          type: string
          required: true
      responses:
        '204':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/notebooks/{id}/ipynbs':
    get:
      description: |
        returns ipynb files on the specified notebook
      operationId: getIPythonNotebooks
      tags:
        - notebook
      parameters:
        - name: id
          in: path
          description: Notebook container ID
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/IPythonNotebook'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Repository:
    description: docker image information
    required:
      - namespace
      - name
    properties:
      namespace:
        description: the repositry's namespace
        type: string
      name:
        description: the image name
        type: string
      description:
        description: the image description
        type: string
  NgcImage:
    description: NGC docker image information
    required:
      - tag
      - size
      - updated
    properties:
      tag:
        description: the image tag
        type: string
      size:
        description: the image size
        type: integer
        format: int64
      updated:
        description: updated unix timestamp
        type: string
        format: date-time
  Image:
    description: docker image information
    required:
      - id
      - repoTags
    properties:
      id:
        description: the image ID
        type: string
      parentID:
        description: the image parent ID
        type: string
      repoDigests:
        description: the image digests
        type: array
        items:
          type: string
      repoTags:
        description: the image name/tag
        type: array
        items:
          type: string
      status:
        description: status of the image
        type: string
      size:
        description: size of the image
        type: integer
        format: int64
      virtualSize:
        description: virtual size of the image
        type: integer
        format: int64
      created:
        description: created unix timestamp
        type: string
  Notebook:
    description: Jupyter notebook information
    required:
      - id
      - name
      - image
    properties:
      id:
        description: the container ID
        type: string
      name:
        description: the container name
        type: string
      image:
        description: the image ID
        type: string
      state:
        description: state of the container
        type: string
      port:
        description: the container published port
        type: integer
        format: int64
      started:
        description: started unix timestamp
        type: string
        format: date-time
  NotebookDetail:
    description: Rescale Job information
    required:
      - id
      - token
    properties:
      id:
        description: the container ID
        type: string
      name:
        description: the container name
        type: string
      state:
        description: state of the container
        type: string
      port:
        description: the container published port
        type: integer
        format: int64
      token:
        description: Jupyter notebook's login token
        type: string
      mounts:
        description: the container labels
        type: array
        items:
          type: string
      started:
        description: started unix timestamp
        type: string
        format: date-time
      ended:
        description: ended unix timestamp
        type: string
        format: date-time
  IPythonNotebook:
    description: .ipynb file attributes
    required:
      - name
    properties:
      name:
        description: file name
        type: string
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: string
      message:
        type: string